// frontend/src/lib/sdk/models/CostCenterResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CostCenterResponse = {
    /**
     * Unique cost center code
     */
    code: string;
    /**
     * Cost center name
     */
    name: string;
    active?: boolean;
    id: number;
    created_at: string;
    updated_at: string;
};